feat: add provisional calls to get project ID for analysis, move local tagged tests to live tests
